requirements.txt
Django==1.10.3
numpy==1.11.2
pandas==0.19.1
python-dateutil==2.6.0
pytz==2016.7
six==1.10.0
xlrd==1.0.0
djangorestframework==3.4.7
django-filter==1.0.0
django-crispy-forms==1.6.1
